Group generic dns-sd machinery in mdns_darwin.go and document the rule
Move browseResult, parseBrowseLine, dnssdBrowseStream, and dnssdBrowse
from discovery_darwin.go to mdns_darwin.go, so all wendy-agnostic
dns-sd browsing and resolution lives in one file and discovery_darwin.go
keeps only the wendy-specific mapping. Add comments on dnssdBrowseStream
and dnssdResolveService stating they must stay free of wendy-specific
logic.
