+++ go/internal/shared/discovery/discovery_darwin.go
@@ -3,11 +3,8 @@
 package discovery
 
 import (
-	"bufio"
 	"context"
 	"fmt"
-	"net"
-	"os/exec"
 	"regexp"
 	"strconv"
 	"strings"
@@ -54,120 +51,6 @@ func discoverLAN(ctx context.Context, timeout time.Duration) ([]models.LANDevice
 	}
 
 	return devices, nil
-}
-
-type browseResult struct {
-	instanceName  string
-	domain        string
-	interfaceName string
-}
-
-// parseBrowseLine parses one line of dns-sd -B output. It returns ok=false
-// for lines that are not "Add" records or are too short to parse.
-func parseBrowseLine(line string) (browseResult, bool) {
-	if !strings.Contains(line, "Add") {
-		return browseResult{}, false
-	}
-	fields := strings.Fields(line)
-	if len(fields) < 7 {
-		return browseResult{}, false
-	}
-	interfaceName := ""
-	if interfaceIndex, err := strconv.Atoi(fields[3]); err == nil {
-		// Silently falls through with empty interfaceName when the
-		// interface disappears between browse and resolve; USB detection
-		// will be skipped for that device rather than returning an error.
-		if iface, ifaceErr := net.InterfaceByIndex(interfaceIndex); ifaceErr == nil {
-			interfaceName = iface.Name
-		}
-	}
-	return browseResult{
-		instanceName:  strings.Join(fields[6:], " "),
-		domain:        fields[4],
-		interfaceName: interfaceName,
-	}, true
-}
-
-// dnssdBrowseStream starts dns-sd -B for serviceType and streams each newly
-// discovered instance to the returned channel. Duplicate (instance, interface)
-// pairs are emitted only once. The channel is closed when ctx is cancelled or
-// the dns-sd process exits; the process is killed and reaped on all paths.
-// Consumers signal they are done by cancelling ctx.
-func dnssdBrowseStream(ctx context.Context, serviceType string) (<-chan browseResult, error) {
-	cmd := exec.CommandContext(ctx, "dns-sd", "-B", serviceType, "local")
-	stdout, err := cmd.StdoutPipe()
-	if err != nil {
-		return nil, err
-	}
-	if err := cmd.Start(); err != nil {
-		return nil, err
-	}
-
-	ch := make(chan browseResult, 16)
-	go func() {
-		defer close(ch)
-		defer func() {
-			_ = cmd.Process.Kill()
-			_ = cmd.Wait()
-		}()
-
-		seen := make(map[string]bool)
-		scanner := bufio.NewScanner(stdout)
-		for scanner.Scan() {
-			result, ok := parseBrowseLine(scanner.Text())
-			if !ok {
-				continue
-			}
-			key := result.instanceName + "%" + result.interfaceName
-			if seen[key] {
-				continue
-			}
-			seen[key] = true
-
-			select {
-			case ch <- result:
-			case <-ctx.Done():
-				return
-			}
-		}
-	}()
-
-	return ch, nil
-}
-
-// dnssdBrowse runs dns-sd -B and returns as soon as results stop arriving.
-// It uses a short settle timer: once the first result arrives, it waits up to
-// 500ms for more results before returning. This avoids waiting for the full timeout.
-func dnssdBrowse(ctx context.Context, serviceType string) ([]browseResult, error) {
-	streamCtx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	browseCh, err := dnssdBrowseStream(streamCtx, serviceType)
-	if err != nil {
-		return nil, err
-	}
-
-	var results []browseResult
-
-	// Wait up to the context deadline for the first result.
-	// Once we get one, use a short settle timer for additional results.
-	var settle <-chan time.Time
-	for {
-		select {
-		case <-ctx.Done():
-			return results, nil
-		case result, open := <-browseCh:
-			if !open {
-				return results, nil
-			}
-			results = append(results, result)
-			// Reset settle timer: wait 500ms for more results.
-			settle = time.After(500 * time.Millisecond)
-		case <-settle:
-			// No new results in 500ms, we're done.
-			return results, nil
-		}
-	}
 }
 
 // dnssdResolve resolves an instance via dns-sd -L and maps the result to a
